[gcd] depart testbench from gcd.scala
