+++ templates/chisel/.gitignore
@@ -15,3 +15,5 @@ target/
 *.fst
 *.fsdb
 *.log
+
+jgproject/
